refactor(storage): split principal-store model modules
